Implemented a Game Class
This class represents a video game in the store's inventory. It stores information about the title of the game, its price, and the quantity available in stock. It has methods to retrieve the title, price, and quantity of the game, as well as a method to update the quantity. This class is used to create and manage game objects within the store's inventory.