Implemented a Customer Class
The customer class represents a customer who can make purchases in the store. It contains information about the customer's funds. Furthermore, the class provides methods to retrieve the customer funds, as well as a method to update the funds. This class is used to create and manage customer objects who interact with the store.